python/pyproject.toml
[project]
name = "python-scripts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-functions>=1.24.0",
    "azure-identity>=1.25.1",
    "azure-storage-queue>=12.15.0",
    "logging>=0.4.9.6",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "ruff>=0.15.0",
    "ty>=0.0.14",
]
